# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "Food Delivery API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get list of restaurants",
                "operationId": "c2732fde08b6fb1b998768e0212d072a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create a new restaurant",
                "operationId": "047e8f41ad48124b86b8af32562c6353",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRestaurantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant created successfully"
                    }
                }
            }
        },
        "/restaurants/all": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get all restaurants with cursor pagination (for bulk fetch)",
                "operationId": "b3a0c5824e366a4473f19f3b4065147b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/restaurants/{restaurant}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get restaurant details",
                "operationId": "be2c8df8f9a7b07410dec1202527f649",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update an existing restaurant",
                "operationId": "a0d1649e226fc5385353d7ce24bb10aa",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRestaurantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant updated successfully"
                    }
                }
            }
        },
        "/restaurants/{restaurant}/operational-status": {
            "patch": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update restaurant operational status",
                "operationId": "00c9e7ce738560133089adf417eab11b",
                "parameters": [
                    {
                        "name": "restaurant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRestaurantOperationalStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant operational status updated successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreRestaurantRequest": {
                "required": [
                    "name",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Warung Nusantara Baru"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deskripsi restoran",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Jalan Merdeka No. 1"
                    },
                    "phone": {
                        "type": "string",
                        "example": "08123456789",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "is_open": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "open_time": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00",
                        "nullable": true
                    },
                    "close_time": {
                        "type": "string",
                        "format": "time",
                        "example": "22:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateRestaurantRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Warung Nusantara Updated"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deskripsi restoran",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "Jalan Merdeka No. 1"
                    },
                    "phone": {
                        "type": "string",
                        "example": "08123456789",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    },
                    "is_open": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "open_time": {
                        "type": "string",
                        "format": "time",
                        "example": "08:00",
                        "nullable": true
                    },
                    "close_time": {
                        "type": "string",
                        "format": "time",
                        "example": "22:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateRestaurantOperationalStatusRequest": {
                "required": [
                    "is_open"
                ],
                "properties": {
                    "is_open": {
                        "type": "boolean",
                        "example": true
                    },
                    "reason": {
                        "type": "string",
                        "example": "Jam operasional",
                        "nullable": true
                    },
                    "changed_by": {
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}